Clear the confusion of where to include the env variables
While I was following the GitHub Actions for deploying on Layer0, I placed the env variables referenced in the workflow to GitHub Secrets. After this didn't work, I reached out to support and it seems like just layer0_deploy_token needs to be added for deploying via CI, but other environment variables need to be added on Layer0 as per the [doc](https://docs.layer0.co/guides/environments#section_accessing_environment_variables_at_build_time).

Creating this pull request to link the sections as this would save lot of time scanning through docs to locate the answer and clear the confusion as to where shall the env. variables be placed during build time.